Backporting cow_ws.erl to 1.0.x from 1.3.0
Backporting cow_ws.erl from 1.3.0 to make use of the permessage-deflate logic (maps
replaced with records for compatability).
An option has been added to configure the takeover and max window bits extensions
along with a compression threshold.
Additionally the zlib deflate stream is now closed completely after use when
no_takeover is set as a memory issue was causing the driver_alloc type to fill up.